Introducing our first annotations
The HibernateCustomerRepositoryImpl has been annotated as a Repository with the @Repository annotation.
Unnecessary methods from previous implementation has been removed from CustomerServiceImpl and we have now annotated that as a @Service.
We have also autowired the application with the @Autowire annotation to pull the HibernateCustomerRepositoryImpl in at the line where we define a customerRepository.